Renvoie de vector de cars

+++ fuzzy/domain/categorie/City.h
@@ -5,7 +5,7 @@
 
 namespace domain {
     class City : public Category{
-        virtual int getType() const { return 1; };
+        virtual double getType() const { return 1; };
     };
 }
